restrucutred helpers functions
