+++ client/src/components/APIClient.ts
@@ -11,8 +11,10 @@ import {
   indicatorsInitialState,
   indicatorsInitialDataStructure,
   IndicatorsDataFromAPI,
+  IndicatorsData,
 } from "./IndicatorsTypes";
 import helpers from "util/helpers";
+import { IndicatorsDatasetId } from "./IndicatorsDatasets";
 
 // API REQUESTS TO THE DATABASE:
 
@@ -36,7 +38,7 @@ function searchForAddressWithGeosearch(q: {
             addrs: [],
             geosearch: undefined,
           });
-        resolve(searchForBBL(splitBBL(firstResult.properties.pad_bbl)));
+        resolve(searchForBBL(helpers.splitBBL(firstResult.properties.pad_bbl)));
       },
       throttleMs: 0,
     });
@@ -44,17 +46,39 @@ function searchForAddressWithGeosearch(q: {
   });
 }
 
-function splitBBL(bbl: string): { boro: string; block: string; lot: string } {
-  return {
-    boro: bbl.slice(0, 1),
-    block: bbl.slice(1, 6),
-    lot: bbl.slice(6, 10),
-  };
+/** Reorganizes raw data from API call and then returns an object that matches the data stucture in state  */
+function createVizData(rawJSON: any, vizType: IndicatorsDatasetId): IndicatorsData {
+  // Generate object to hold data for viz
+  // Note: keys in "values" object need to match exact key names in data from API call
+  var vizData: IndicatorsData = Object.assign({}, indicatorsInitialDataStructure[vizType]);
+
+  vizData.labels = [];
+  for (const column in vizData.values) {
+    vizData.values[column] = [];
+  }
+
+  // Generate arrays of data for chart.js visualizations:
+  // Default grouping is by MONTH
+
+  const rawJSONLength = rawJSON.length;
+
+  for (let i = 0; i < rawJSONLength; i++) {
+    vizData.labels.push(rawJSON[i].month);
+
+    for (const column in vizData.values) {
+      const vizTypePlusColumn = vizType + "_" + column;
+      const values = vizData.values[column];
+      if (!values)
+        throw new Error(`Column "${column}" of visualization "${vizType}" is not an array!`);
+      values.push(parseInt(rawJSON[i][vizTypePlusColumn]));
+    }
+  }
+  return vizData;
 }
 
 function searchForAddress(q: SearchAddress): Promise<SearchResults> {
   if (q.bbl) {
-    return searchForBBL(splitBBL(q.bbl));
+    return searchForBBL(helpers.splitBBL(q.bbl));
   }
   return searchForAddressWithGeosearch(q);
 }
@@ -79,7 +103,7 @@ async function getIndicatorHistory(bbl: string): Promise<IndicatorsDataFromAPI> 
   const structuredIndicatorData = Object.assign({}, indicatorsInitialDataStructure);
 
   for (const indicator of indicatorsInitialState.indicatorList) {
-    var inputData = helpers.createVizData(rawIndicatorData, indicator);
+    var inputData = createVizData(rawIndicatorData, indicator);
     // TO DO: Fix this "any" typecasting
     structuredIndicatorData[indicator] = inputData as any;
   }
